refactor tests for $$.sizeGreaterThanOrEqual
replaced useless unit-tests by analogous integration test
