+++ src/test/java/integration/collections/CollectionMethodsTest.java
@@ -403,10 +403,6 @@ final class CollectionMethodsTest extends ITest {
       .as(description, "size").isInstanceOf(ListSizeMismatch.class)
       .hasCauseExactlyInstanceOf(NoSuchElementException.class);
 
-    assertThatThrownBy(() -> elementsCollection.shouldHave(sizeGreaterThanOrEqual(1)))
-      .as(description, "sizeGreaterThanOrEqual").isInstanceOf(ListSizeMismatch.class)
-      .hasCauseExactlyInstanceOf(NoSuchElementException.class);
-
     assertThatThrownBy(() -> elementsCollection.shouldHave(sizeLessThan(0)))
       .as(description, "sizeLessThan").isInstanceOf(ListSizeMismatch.class)
       .hasCauseExactlyInstanceOf(NoSuchElementException.class);
@@ -435,10 +431,6 @@ final class CollectionMethodsTest extends ITest {
 
     assertThatThrownBy(() -> elementsCollection.shouldHave(size(1)))
       .as(description, "size").isInstanceOf(ElementNotFound.class)
-      .hasCauseExactlyInstanceOf(IndexOutOfBoundsException.class);
-
-    assertThatThrownBy(() -> elementsCollection.shouldHave(sizeGreaterThanOrEqual(1)))
-      .as(description, "sizeGreaterThanOrEqual").isInstanceOf(ElementNotFound.class)
       .hasCauseExactlyInstanceOf(IndexOutOfBoundsException.class);
 
     assertThatThrownBy(() -> elementsCollection.shouldHave(sizeLessThan(0)))
